requirements.txt
# Core ML/RL libraries
torch>=2.0.0
stable-baselines3>=2.0.0
gymnasium>=0.29.0
numpy>=1.24.0
pandas>=2.0.0

# LLM for event extraction
transformers>=4.35.0
accelerate>=0.24.0
# Note: Qwen models are loaded via transformers from HuggingFace

# Market data
yfinance>=0.2.0
pandas-datareader>=0.10.0
# Note: Stooq (stooq.com) is a data source website, not a Python package
# Data can be downloaded manually from https://stooq.com/db/h/ if needed

# Data processing
scikit-learn>=1.3.0
scipy>=1.11.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
python-dotenv>=1.0.0
wandb>=0.15.0  # Optional: for experiment tracking

# Evaluation metrics
quantstats>=0.0.62
